+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Uncaught exception which prevented discovering
+  non-decryptable members.
+
 ## [2.3.2] - 2021-04-22
 
 ### Fixed